test: introduce fixtures module
We have a lot of code duplciation in fuzzers when creating storage,
grid, superblock, tracer, etc. This is bad, as it makes refactoring
construcotrs unnecessary painful (unnecessary, because typically there's
only _producion_ call-site, all duplicates come from fuzzers).

The idea is that everything should be create through fixtures module, so
that there's just a single consturciton call, shared for all fuzzer.
Crucially, the fixtures can _default_ most of the arguments, and
otherwise adapt the API, to insulate callers from changes.